backend/requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
pandas==2.1.1
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
aiohttp==3.9.3
sqlalchemy==2.0.22
pydantic==2.4.2
groq==0.4.2
